Add installation instructions, RELEASE!
N.B. a lot of things are broken right now, but they're (as far as
we can tell) broken in a non-distruptive way, i.e. the new behavior
doesn't work and it just reverts to the underlying BTW logic.  That
is probably good enough for a first beta/revival release.

The documentation is all "broken" right now, and any case where the
code for a documented feature doesn't work can be treated as the
documentation just being wrong.